improve documentation of web and cmd packages
