+++ cmd/README.md
@@ -1,7 +1,3 @@
-# cmd
-
-The command line interface for the application.
-
 <!-- gomarkdoc:embed:start -->
 
 <!-- Code generated by gomarkdoc. DO NOT EDIT -->
@@ -36,7 +32,7 @@ The main packages are:
 
 
 <a name="AddRoutes"></a>
-## func AddRoutes
+## func [AddRoutes](<https://github.com/conneroisu/steroscopic-hardware/blob/main/cmd/routes.go#L42-L48>)
 
 ```go
 func AddRoutes(mux *http.ServeMux, logger *logger.Logger, params *despair.Parameters, leftStream, rightStream, outputStream *camera.StreamManager, cancel context.CancelFunc) error
@@ -67,7 +63,7 @@ Parameters:
 Returns any error encountered during route configuration.
 
 <a name="NewServer"></a>
-## func NewServer
+## func [NewServer](<https://github.com/conneroisu/steroscopic-hardware/blob/main/cmd/root.go#L63-L68>)
 
 ```go
 func NewServer(logger *logger.Logger, params *despair.Parameters, leftStream, rightStream, outputStream *camera.StreamManager, cancel context.CancelFunc) (http.Handler, error)
@@ -89,7 +85,7 @@ Parameters:
 Returns an http.Handler and any error encountered during setup.
 
 <a name="Run"></a>
-## func Run
+## func [Run](<https://github.com/conneroisu/steroscopic-hardware/blob/main/cmd/root.go#L110-L113>)
 
 ```go
 func Run(ctx context.Context, onStart func()) error
