+++ cmd/handlers/README.md
@@ -25,8 +25,8 @@ Package handlers contains functions for handling API requests.
   - [func ParametersHandler\(logger \*logger.Logger, params \*despair.Parameters\) APIFn](<#ParametersHandler>)
   - [func StreamHandlerFn\(manager \*camera.StreamManager\) APIFn](<#StreamHandlerFn>)
 
-<a name="Make"></a>
 
+<a name="Make"></a>
 ## func [Make](<https://github.com/conneroisu/steroscopic-hardware/blob/main/cmd/handlers/api.go#L15>)
 
 ```go
@@ -36,7 +36,6 @@ func Make(fn APIFn) http.HandlerFunc
 Make returns a function that can be used as an http.HandlerFunc.
 
 <a name="MorphableHandler"></a>
-
 ## func [MorphableHandler](<https://github.com/conneroisu/steroscopic-hardware/blob/main/cmd/handlers/api.go#L34-L37>)
 
 ```go
@@ -46,7 +45,6 @@ func MorphableHandler(wrapper func(templ.Component) templ.Component, morph templ
 MorphableHandler returns a handler that checks for the presence of the hx\-trigger header and serves either the full or morphed view.
 
 <a name="APIFn"></a>
-
 ## type [APIFn](<https://github.com/conneroisu/steroscopic-hardware/blob/main/cmd/handlers/api.go#L12>)
 
 APIFn is a function that handles an API request.
@@ -56,7 +54,6 @@ type APIFn func(w http.ResponseWriter, r *http.Request) error
 ```
 
 <a name="ConfigureCamera"></a>
-
 ### func [ConfigureCamera](<https://github.com/conneroisu/steroscopic-hardware/blob/main/cmd/handlers/configure.go#L58-L63>)
 
 ```go
@@ -66,7 +63,6 @@ func ConfigureCamera(logger *logger.Logger, params *despair.Parameters, leftStre
 ConfigureCamera handles client requests to configure all camera parameters at once.
 
 <a name="ErrorHandler"></a>
-
 ### func [ErrorHandler](<https://github.com/conneroisu/steroscopic-hardware/blob/main/cmd/handlers/api.go#L49-L51>)
 
 ```go
@@ -76,7 +72,6 @@ func ErrorHandler(fn APIFn) APIFn
 ErrorHandler returns a handler that returns an error response.
 
 <a name="GetPorts"></a>
-
 ### func [GetPorts](<https://github.com/conneroisu/steroscopic-hardware/blob/main/cmd/handlers/ports.go#L15-L17>)
 
 ```go
@@ -86,7 +81,6 @@ func GetPorts(logger *logger.Logger) APIFn
 GetPorts handles client requests to configure the camera.
 
 <a name="ParametersHandler"></a>
-
 ### func [ParametersHandler](<https://github.com/conneroisu/steroscopic-hardware/blob/main/cmd/handlers/configure.go#L15>)
 
 ```go
@@ -96,7 +90,6 @@ func ParametersHandler(logger *logger.Logger, params *despair.Parameters) APIFn
 ParametersHandler handles client requests to change the parameters of the desparity map generator.
 
 <a name="StreamHandlerFn"></a>
-
 ### func [StreamHandlerFn](<https://github.com/conneroisu/steroscopic-hardware/blob/main/cmd/handlers/stream.go#L17>)
 
 ```go
@@ -106,5 +99,6 @@ func StreamHandlerFn(manager *camera.StreamManager) APIFn
 StreamHandlerFn returns a handler for streaming camera images to multiple clients
 
 Generated by [gomarkdoc](<https://github.com/princjef/gomarkdoc>)
+
 
 <!-- gomarkdoc:embed:end -->
