+++ pkg/web/README.md
@@ -1,0 +1,105 @@
+<!-- gomarkdoc:embed:start -->
+
+<!-- Code generated by gomarkdoc. DO NOT EDIT -->
+
+# web
+
+```go
+import "github.com/conneroisu/steroscopic-hardware/pkg/web"
+```
+
+Package web contains SVG templates and dom targets for the web UI.
+
+SVG templates are used to render SVG icons and text in the web UI. Templates are embedded into the package using the go:embed directive.
+
+Targets are used to identify DOM elements in the web UI and are defined as constants with a unique ID and a CSS selector. Targets are used to update the DOM with new content or to insert new elements.
+
+## Index
+
+- [Variables](<#variables>)
+- [type Target](<#Target>)
+
+
+## Variables
+
+<a name="TargetLogContainer"></a>
+
+```go
+var (
+    // TargetLogContainer is a target for the log container.
+    // It is used to insert log entries into the DOM.
+    TargetLogContainer = Target{
+        ID:  "log-container",
+        Sel: "#log-container",
+    }
+
+    // TargetStatusContent is a target for the status content.
+    // It is used to update the status content in the DOM.
+    TargetStatusContent = Target{
+        ID:  "status-content",
+        Sel: "#status-content",
+    }
+)
+```
+
+<a name="CircleQuestion"></a>CircleQuestion is a template for the SVG circle\-question icon.
+
+```go
+var CircleQuestion = templ.Raw(circleQuestion)
+```
+
+<a name="CircleX"></a>CircleX is a template for the SVG circle\-x icon.
+
+```go
+var CircleX = templ.Raw(circleX)
+```
+
+<a name="GreenUp"></a>GreenUp is a template for the SVG green\-up icon.
+
+```go
+var GreenUp = templ.Raw(greenUp)
+```
+
+<a name="LivePageTitle"></a>
+
+```go
+var (
+    // LivePageTitle is the title of the live page.
+    LivePageTitle = "Live Camera System"
+)
+```
+
+<a name="RedDown"></a>RedDown is a template for the SVG red\-down icon.
+
+```go
+var RedDown = templ.Raw(redDown)
+```
+
+<a name="RefreshCw"></a>RefreshCw is a template for the SVG refresh\-cw icon.
+
+```go
+var RefreshCw = templ.Raw(refreshCw)
+```
+
+<a name="SettingsGear"></a>SettingsGear is a template for the SVG settings\-geat icon.
+
+```go
+var SettingsGear = templ.Raw(settingsGear)
+```
+
+<a name="Target"></a>
+## type [Target](<https://github.com/conneroisu/steroscopic-hardware/blob/main/pkg/web/targets.go#L4-L7>)
+
+Target is a struct representing a dom target.
+
+```go
+type Target struct {
+    ID  string `json:"id"`
+    Sel string `json:"sel"`
+}
+```
+
+Generated by [gomarkdoc](<https://github.com/princjef/gomarkdoc>)
+
+
+<!-- gomarkdoc:embed:end -->
